refactor to allow putting assignment info on internal events
