+++ rust/.sqlx/query-2a1514685491269ca61a0a898d7491f908b5047e86b13fbff22cf8f58e85cc81.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "\n            SELECT id, team_id, status, name, description, false as \"eligible_for_assignment!\" FROM posthog_errortrackingissue\n            WHERE team_id = $1 AND id = $2\n            ",
+  "query": "\n            SELECT id, team_id, status, name, description FROM posthog_errortrackingissue\n            WHERE team_id = $1 AND id = $2\n            ",
   "describe": {
     "columns": [
       {
@@ -27,11 +27,6 @@
         "ordinal": 4,
         "name": "description",
         "type_info": "Text"
-      },
-      {
-        "ordinal": 5,
-        "name": "eligible_for_assignment!",
-        "type_info": "Bool"
       }
     ],
     "parameters": {
@@ -45,9 +40,8 @@
       false,
       false,
       true,
-      true,
-      null
+      true
     ]
   },
-  "hash": "8a64e4bb0150cfafd437bd85da3bb375e3b82d5679424c479518e0ed0dbe850a"
+  "hash": "2a1514685491269ca61a0a898d7491f908b5047e86b13fbff22cf8f58e85cc81"
 }
