--- rust/.sqlx/query-5001fc5e1fc02ac57e886cc765f4d0a3cbc2086d1c8ad135f22ed87b9bea65e2.json
@@ -1,53 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            -- the \"eligible_for_assignment!\" forces sqlx to assume not null, which is correct in this case, but\n            -- generally a risky override of sqlx's normal type checking\n            SELECT i.id, i.team_id, i.status, i.name, i.description, false as \"eligible_for_assignment!\"\n            FROM posthog_errortrackingissue i\n            JOIN posthog_errortrackingissuefingerprintv2 f ON i.id = f.issue_id\n            WHERE f.team_id = $1 AND f.fingerprint = $2\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 1,
-        "name": "team_id",
-        "type_info": "Int4"
-      },
-      {
-        "ordinal": 2,
-        "name": "status",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 3,
-        "name": "name",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 4,
-        "name": "description",
-        "type_info": "Text"
-      },
-      {
-        "ordinal": 5,
-        "name": "eligible_for_assignment!",
-        "type_info": "Bool"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Int4",
-        "Text"
-      ]
-    },
-    "nullable": [
-      false,
-      false,
-      false,
-      true,
-      true,
-      null
-    ]
-  },
-  "hash": "5001fc5e1fc02ac57e886cc765f4d0a3cbc2086d1c8ad135f22ed87b9bea65e2"
-}
