+++ rust/.sqlx/query-8029a54580dbb1cbdead38b30aba5c7b0a9fb1ea7f3bb89f67222b2d1f18be6f.json
@@ -1,0 +1,47 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            -- the \"eligible_for_assignment!\" forces sqlx to assume not null, which is correct in this case, but\n            -- generally a risky override of sqlx's normal type checking\n            SELECT i.id, i.team_id, i.status, i.name, i.description\n            FROM posthog_errortrackingissue i\n            JOIN posthog_errortrackingissuefingerprintv2 f ON i.id = f.issue_id\n            WHERE f.team_id = $1 AND f.fingerprint = $2\n            ",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 1,
+        "name": "team_id",
+        "type_info": "Int4"
+      },
+      {
+        "ordinal": 2,
+        "name": "status",
+        "type_info": "Text"
+      },
+      {
+        "ordinal": 3,
+        "name": "name",
+        "type_info": "Text"
+      },
+      {
+        "ordinal": 4,
+        "name": "description",
+        "type_info": "Text"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Int4",
+        "Text"
+      ]
+    },
+    "nullable": [
+      false,
+      false,
+      false,
+      true,
+      true
+    ]
+  },
+  "hash": "8029a54580dbb1cbdead38b30aba5c7b0a9fb1ea7f3bb89f67222b2d1f18be6f"
+}
