+++ rust/.sqlx/query-9988d1afe606a68be208a99cdec5bea052b9445ce656fc9d8ef0215663dcfe96.json
@@ -1,0 +1,49 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n                INSERT INTO posthog_errortrackingissueassignment (id, issue_id, user_id, user_group_id, created_at)\n                VALUES ($1, $2, $3, $4, NOW())\n                ON CONFLICT (issue_id) DO UPDATE SET issue_id = $2 -- no-op to get a returned row\n                RETURNING id, issue_id, user_id, user_group_id, created_at\n            ",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 1,
+        "name": "issue_id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 2,
+        "name": "user_id",
+        "type_info": "Int4"
+      },
+      {
+        "ordinal": 3,
+        "name": "user_group_id",
+        "type_info": "Uuid"
+      },
+      {
+        "ordinal": 4,
+        "name": "created_at",
+        "type_info": "Timestamptz"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Uuid",
+        "Uuid",
+        "Int4",
+        "Uuid"
+      ]
+    },
+    "nullable": [
+      false,
+      false,
+      true,
+      true,
+      false
+    ]
+  },
+  "hash": "9988d1afe606a68be208a99cdec5bea052b9445ce656fc9d8ef0215663dcfe96"
+}
