--- rust/.sqlx/query-dd8181336020206bab40acaa11ed60dd352acef34d7b56cf79b890d2c86f90d6.json
@@ -1,17 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n                INSERT INTO posthog_errortrackingissueassignment (id, issue_id, user_id, user_group_id, created_at)\n                VALUES ($1, $2, $3, $4, NOW())\n                ON CONFLICT (issue_id) DO NOTHING\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Uuid",
-        "Int4",
-        "Uuid"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "dd8181336020206bab40acaa11ed60dd352acef34d7b56cf79b890d2c86f90d6"
-}
