+++ rust/cymbal/src/issue_resolution.rs
@@ -1,4 +1,5 @@
 use std::fmt::Display;
+use std::sync::Arc;
 
 use chrono::{DateTime, Utc};
 use common_kafka::kafka_messages::internal_events::{InternalEvent, InternalEventEvent};
@@ -7,6 +8,8 @@ use common_kafka::kafka_producer::send_iter_to_kafka;
 use sqlx::Acquire;
 use uuid::Uuid;
 
+use crate::assignment_rules::{assign_issue, Assignment};
+use crate::types::FingerprintedErrProps;
 use crate::{
     app_context::AppContext,
     error::UnhandledError,
@@ -30,7 +33,6 @@ pub struct Issue {
     pub status: IssueStatus,
     pub name: Option<String>,
     pub description: Option<String>,
-    pub eligible_for_assignment: bool,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -56,7 +58,7 @@ impl Issue {
             r#"
             -- the "eligible_for_assignment!" forces sqlx to assume not null, which is correct in this case, but
             -- generally a risky override of sqlx's normal type checking
-            SELECT i.id, i.team_id, i.status, i.name, i.description, false as "eligible_for_assignment!"
+            SELECT i.id, i.team_id, i.status, i.name, i.description
             FROM posthog_errortrackingissue i
             JOIN posthog_errortrackingissuefingerprintv2 f ON i.id = f.issue_id
             WHERE f.team_id = $1 AND f.fingerprint = $2
@@ -81,7 +83,7 @@ impl Issue {
         let res = sqlx::query_as!(
             Issue,
             r#"
-            SELECT id, team_id, status, name, description, false as "eligible_for_assignment!" FROM posthog_errortrackingissue
+            SELECT id, team_id, status, name, description FROM posthog_errortrackingissue
             WHERE team_id = $1 AND id = $2
             "#,
             team_id,
@@ -108,7 +110,6 @@ impl Issue {
             status: IssueStatus::Active,
             name: Some(name),
             description: Some(description),
-            eligible_for_assignment: true,
         };
 
         sqlx::query!(
@@ -128,17 +129,13 @@ impl Issue {
         Ok(issue)
     }
 
-    pub async fn maybe_reopen<'c, E>(
-        &mut self,
-        executor: E,
-        context: &AppContext,
-    ) -> Result<(), UnhandledError>
+    pub async fn maybe_reopen<'c, E>(&mut self, executor: E) -> Result<bool, UnhandledError>
     where
         E: sqlx::Executor<'c, Database = sqlx::Postgres>,
     {
         // If this issue is already active, or permanently suppressed, we don't need to do anything
         if matches!(self.status, IssueStatus::Active | IssueStatus::Suppressed) {
-            return Ok(());
+            return Ok(false);
         }
 
         let res = sqlx::query_scalar!(
@@ -153,16 +150,28 @@ impl Issue {
         .fetch_all(executor)
         .await?;
 
-        self.eligible_for_assignment = !res.is_empty(); // We actually updated a row
+        Ok(!res.is_empty()) // We actually updated a row
+    }
 
-        // If we actually updated a row
-        if self.eligible_for_assignment {
-            metrics::counter!(ISSUE_REOPENED).increment(1);
-            capture_issue_reopened(self.team_id, self.id);
-            send_issue_reopened_alert(context, self).await?;
-        }
+    pub async fn get_assignments<'c, E>(
+        &self,
+        executor: E,
+    ) -> Result<Vec<Assignment>, UnhandledError>
+    where
+        E: sqlx::Executor<'c, Database = sqlx::Postgres>,
+    {
+        let assignments = sqlx::query_as!(
+            Assignment,
+            r#"
+            SELECT id, issue_id, user_id, user_group_id, created_at FROM posthog_errortrackingissueassignment
+            WHERE issue_id = $1
+            "#,
+            self.id
+        )
+        .fetch_all(executor)
+        .await?;
 
-        Ok(())
+        Ok(assignments)
     }
 }
 
@@ -220,19 +229,29 @@ impl IssueFingerprintOverride {
 }
 
 pub async fn resolve_issue(
-    context: &AppContext,
+    context: Arc<AppContext>,
     team_id: i32,
-    fingerprint: &str,
     name: String,
     description: String,
     event_timestamp: DateTime<Utc>,
+    event_properties: FingerprintedErrProps,
 ) -> Result<Issue, UnhandledError> {
+    let fingerprint = &event_properties.fingerprint;
     let mut conn = context.pool.acquire().await?;
-
     // Fast path - just fetch the issue directly, and then reopen it if needed
     let existing_issue = Issue::load_by_fingerprint(&mut *conn, team_id, fingerprint).await?;
     if let Some(mut issue) = existing_issue {
-        issue.maybe_reopen(&mut *conn, context).await?;
+        if issue.maybe_reopen(&mut *conn).await? {
+            metrics::counter!(ISSUE_REOPENED).increment(1);
+            capture_issue_reopened(issue.team_id, issue.id);
+            let assignment = assign_issue(
+                context.clone(),
+                issue.clone(),
+                event_properties.to_output(issue.id),
+            )
+            .await?;
+            send_issue_reopened_alert(&context, &issue, assignment).await?;
+        }
         return Ok(issue);
     }
 
@@ -265,25 +284,19 @@ pub async fn resolve_issue(
     // saving both the issue and the override. Otherwise, rollback the transaction, and
     // use the retrieved issue override.
     let was_created = issue_override.issue_id == issue.id;
-    let mut issue = issue;
     if !was_created {
-        // It wasn't actually created new, so it's not eligible for assignment based
-        // on it being new - maybe_reopen will handle eligibility of the existing issue
-        issue.eligible_for_assignment = false;
         txn.rollback().await?;
     } else {
         metrics::counter!(ISSUE_CREATED).increment(1);
-        send_issue_created_alert(context, &issue).await?;
+        let assignment = assign_issue(
+            context.clone(),
+            issue.clone(),
+            event_properties.to_output(issue.id),
+        )
+        .await?;
+        send_issue_created_alert(&context, &issue, assignment).await?;
         txn.commit().await?;
         capture_issue_created(team_id, issue_override.issue_id);
-    }
-
-    // This being None is /almost/ impossible, unless between the transaction above finishing and
-    // this point, someone merged the issue and deleted the old one, but if that happens,
-    // we don't care about this reopen failing (since this issue is irrelevant anyway).
-    if let Some(mut loaded) = Issue::load(&mut *conn, team_id, issue_override.issue_id).await? {
-        loaded.maybe_reopen(&mut *conn, context).await?;
-        issue = loaded;
     }
 
     Ok(issue)
@@ -292,21 +305,24 @@ pub async fn resolve_issue(
 async fn send_issue_created_alert(
     context: &AppContext,
     issue: &Issue,
+    assignment: Option<Assignment>,
 ) -> Result<(), UnhandledError> {
-    send_internal_event(context, "$error_tracking_issue_created", issue).await
+    send_internal_event(context, "$error_tracking_issue_created", issue, assignment).await
 }
 
 async fn send_issue_reopened_alert(
     context: &AppContext,
     issue: &Issue,
+    assignment: Option<Assignment>,
 ) -> Result<(), UnhandledError> {
-    send_internal_event(context, "$error_tracking_issue_reopened", issue).await
+    send_internal_event(context, "$error_tracking_issue_reopened", issue, assignment).await
 }
 
 async fn send_internal_event(
     context: &AppContext,
     event: &str,
     issue: &Issue,
+    new_assignment: Option<Assignment>,
 ) -> Result<(), UnhandledError> {
     let mut event = InternalEventEvent::new(event, issue.id, Utc::now(), None);
     event
@@ -315,6 +331,19 @@ async fn send_internal_event(
     event
         .insert_prop("description", issue.description.clone())
         .expect("Strings are serializable");
+
+    if let Some(assignment) = new_assignment {
+        if let Some(user_id) = assignment.user_id {
+            event
+                .insert_prop("assigned_to", user_id.to_string())
+                .expect("Strings are serializable");
+        }
+        if let Some(group_id) = assignment.user_group_id {
+            event
+                .insert_prop("assigned_to_team", group_id.to_string())
+                .expect("Strings are serializable");
+        }
+    }
 
     send_iter_to_kafka(
         &context.immediate_producer,
